+++ Changelog.md
@@ -9,8 +9,11 @@ __New Features__
     - **Breaking Change**: `OutdoorTemperature` values must be 82F and 95F (plus an optional user-specified temperature >95F) for cooling
     - **Breaking Change**: Additional datapoints with `CapacityDescription`="nominal" must be provided for variable speed equipment
     - Detailed performance datapoints can now be used for single stage and two stage equipment too.
-- Allows detailed modeling of electric vehicles (batteries and charging/discharging) using `Vehicles` as an alternative to the simple EV charging plug load.
+- Allows detailed modeling of electric vehicles (batteries and charging/discharging) using `Vehicles` as an alternative to the simple EV charging `PlugLoad`.
 - Allows requesting timeseries EnergyPlus output meters (e.g., `--hourly "MainsWater:Facility"`), similar to requesting EnergyPlus output variables.
+- BuildResidentialScheduleFile measure:
+  - Adds stochastic schedule generation for electric vehicle charging.
+  - Removes generation of stochastic schedules for building components not present in the HPXML file.
 
 __Bugfixes__
 - Fixes zero occupants specified for one unit in a whole MF building from being treated like zero occupants for every unit.